pyproject.toml
[project]
name = "soter-market"
version = "0.1.0"
description = "Soter Market"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=2.22.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "celery>=5.5.2",
    "fastapi-cache2[redis]>=0.2.2",
    "fastapi[standard]>=0.115.12",
    "flower>=2.0.1",
    "greenlet>=3.2.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-jose>=3.4.0",
    "python-json-logger>=3.3.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
]
